+++ src/components/__tests__/Icon.test.tsx
@@ -1,0 +1,16 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+import React from 'react';
+
+import Icon from '../Icon';
+
+import renderWithProviders from '../../../test/renderWithProviders';
+
+describe('Icon component', () => {
+  it('renders correctly', () => {
+    const { container } = renderWithProviders(<Icon icon="tune" />, 'fi');
+
+    expect(container).toMatchSnapshot();
+  });
+});