+++ src/components/form/RegisterPersonForm.tsx
@@ -2,12 +2,12 @@
 // Licensed under the MIT License
 
 import React from 'react';
-import styled, { css } from 'styled-components';
-import { ArrowLeftIcon, ArrowRightIcon, PaperAirplaneIcon } from '@primer/octicons-react';
+import styled from 'styled-components';
 import { IntlShape, injectIntl } from 'react-intl';
 
 import Button from '../Button';
 import FormDiv from './FormDiv';
+import Icon from '../Icon';
 import LocalizedLink from '../link/LocalizedLink';
 import RadioInput from './RadioInput';
 import SwitchCheckbox from './SwitchCheckbox';
@@ -33,22 +33,6 @@ const FormContainer = styled.div`
 const FormPage = styled.div``;
 
 const ButtonDiv = styled.div``;
-
-const iconStyle = css`
-  margin: 0;
-`;
-
-const ArrowLeft = styled(ArrowLeftIcon)`
-  ${iconStyle}
-`;
-
-const ArrowRight = styled(ArrowRightIcon)`
-  ${iconStyle}
-`;
-
-const PaperAirplane = styled(PaperAirplaneIcon)`
-  ${iconStyle}
-`;
 
 type Props = {
   intl: IntlShape;
@@ -620,17 +604,17 @@ class RegisterPersonForm extends React.Component<Props, State> {
           <FormDiv>
             <ButtonDiv hidden={currentPage === 0}>
               <Button onClick={this.moveToPreviousPage}>
-                <ArrowLeft size={24} /> <span>{intl('clientRegisterPersonFormPrevious')}</span>
+                <Icon icon="arrow_back" /> <span>{intl('clientRegisterPersonFormPrevious')}</span>
               </Button>
             </ButtonDiv>
             <ButtonDiv hidden={currentPage === 2}>
               <Button onClick={this.moveToNextPage}>
-                <span>{intl('clientRegisterPersonFormNext')}</span> <ArrowRight size={24} />
+                <span>{intl('clientRegisterPersonFormNext')}</span> <Icon icon="arrow_forward" />
               </Button>
             </ButtonDiv>
             <ButtonDiv hidden={currentPage !== 2}>
               <button type="submit">
-                <PaperAirplane size={24} /> {intl('clientRegisterPersonFormSend')}
+                <Icon icon="send" /> {intl('clientRegisterPersonFormSend')}
               </button>
             </ButtonDiv>
           </FormDiv>