+++ src/components/search/SearchForm.tsx
@@ -4,8 +4,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled, { css } from 'styled-components';
-import { SearchIcon } from '@primer/octicons-react';
 import { useIntl } from 'react-intl';
+
+import Icon from '../Icon';
 
 import createInternationalization from '../../util/createInternationalization';
 
@@ -45,12 +46,8 @@ const Input = styled.input`
   ${open}
 `;
 
-const Icon = styled(SearchIcon)`
-  width: 1em;
-  height: auto;
-  margin: 0.3em;
-  color: var(--color-text);
-  pointer-events: none;
+const MaterialIcon = styled(Icon)`
+  margin: 0 6px;
 `;
 
 const propTypes = {
@@ -77,7 +74,7 @@ function SearchForm({ error, loading, onFocus, searchData, searchQuery }) {
           type="text"
           value={searchQuery}
         />
-        <Icon />
+        <MaterialIcon icon="search" />
       </Form>
     );
   }
@@ -91,7 +88,7 @@ function SearchForm({ error, loading, onFocus, searchData, searchQuery }) {
         type="text"
         value={searchQuery}
       />
-      <Icon />
+      <MaterialIcon icon="search" />
     </Form>
   );
 }