+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.30.0] - 2021-09-12
+
+### Added
+
+- Material icon component.
+- Icons to cookie settings.
+
+### Changed
+
+- Icons on the site to be Material icons.
+
+### Removed
+
+- Octicons from the site.
+
+### Fixed
+
+- Name of the default social media image in the `head`.
+
 ## [0.29.2] - 2021-08-16
 
 ### Fixed
@@ -776,7 +795,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.29.2...HEAD
+[unreleased]: https://github.com/visiosto/visiosto-website/compare/v0.30.0...HEAD
+[0.30.0]: https://github.com/visiosto/visiosto-website/compare/v0.29.2...v0.30.0
 [0.29.2]: https://github.com/visiosto/visiosto-website/compare/v0.29.1...v0.29.2
 [0.29.1]: https://github.com/visiosto/visiosto-website/compare/v0.29.0...v0.29.1
 [0.29.0]: https://github.com/visiosto/visiosto-website/compare/v0.28.0...v0.29.0